Added player dead sprite.
Changes to be committed:
	renamed:    characters/mob.cs -> characters/mob/mob.cs
	renamed:    characters/mob.tscn -> characters/mob/mob.tscn
	renamed:    art/enemyFlyingAlt_1.png -> characters/mob/sprite/enemyFlyingAlt_1.png
	renamed:    art/enemyFlyingAlt_1.png.import -> characters/mob/sprite/enemyFlyingAlt_1.png.import
	renamed:    art/enemyFlyingAlt_2.png -> characters/mob/sprite/enemyFlyingAlt_2.png
	renamed:    art/enemyFlyingAlt_2.png.import -> characters/mob/sprite/enemyFlyingAlt_2.png.import
	renamed:    art/enemySwimming_1.png -> characters/mob/sprite/enemySwimming_1.png
	renamed:    art/enemySwimming_1.png.import -> characters/mob/sprite/enemySwimming_1.png.import
	renamed:    art/enemySwimming_2.png -> characters/mob/sprite/enemySwimming_2.png
	renamed:    art/enemySwimming_2.png.import -> characters/mob/sprite/enemySwimming_2.png.import
	renamed:    art/enemyWalking_1.png -> characters/mob/sprite/enemyWalking_1.png
	renamed:    art/enemyWalking_1.png.import -> characters/mob/sprite/enemyWalking_1.png.import
	renamed:    art/enemyWalking_2.png -> characters/mob/sprite/enemyWalking_2.png
	renamed:    art/enemyWalking_2.png.import -> characters/mob/sprite/enemyWalking_2.png.import
	renamed:    characters/player.cs -> characters/player/player.cs
	renamed:    characters/player.tscn -> characters/player/player.tscn
	new file:   characters/player/sprite/playerGrey_dead1.png
	new file:   characters/player/sprite/playerGrey_dead1.png.import
	renamed:    art/playerGrey_up1.png -> characters/player/sprite/playerGrey_up1.png
	renamed:    art/playerGrey_up1.png.import -> characters/player/sprite/playerGrey_up1.png.import
	renamed:    art/playerGrey_up2.png -> characters/player/sprite/playerGrey_up2.png
	renamed:    art/playerGrey_up2.png.import -> characters/player/sprite/playerGrey_up2.png.import
	renamed:    art/playerGrey_walk1.png -> characters/player/sprite/playerGrey_walk1.png
	renamed:    art/playerGrey_walk1.png.import -> characters/player/sprite/playerGrey_walk1.png.import
	renamed:    art/playerGrey_walk2.png -> characters/player/sprite/playerGrey_walk2.png
	renamed:    art/playerGrey_walk2.png.import -> characters/player/sprite/playerGrey_walk2.png.import
	modified:   export_presets.cfg
	modified:   main.cs
	modified:   main.tscn
