+++ characters/mob/sprite/enemyFlyingAlt_1.png
