+++ characters/mob/sprite/enemyFlyingAlt_1.png.import
@@ -2,16 +2,16 @@
 
 importer="texture"
 type="CompressedTexture2D"
-uid="uid://bjrr3o2ld1wpi"
-path="res://.godot/imported/enemySwimming_1.png-dd0e11759dc3d624c8a704f6e98a3d80.ctex"
+uid="uid://cf7vccoi7pb40"
+path="res://.godot/imported/enemyFlyingAlt_1.png-01f6fd67cf9ef2eff5afcbabf2608d42.ctex"
 metadata={
 "vram_texture": false
 }
 
 [deps]
 
-source_file="res://art/enemySwimming_1.png"
-dest_files=["res://.godot/imported/enemySwimming_1.png-dd0e11759dc3d624c8a704f6e98a3d80.ctex"]
+source_file="res://characters/mob/sprite/enemyFlyingAlt_1.png"
+dest_files=["res://.godot/imported/enemyFlyingAlt_1.png-01f6fd67cf9ef2eff5afcbabf2608d42.ctex"]
 
 [params]
