+++ characters/mob/sprite/enemyFlyingAlt_2.png
