+++ characters/mob/sprite/enemyFlyingAlt_2.png.import
@@ -2,16 +2,16 @@
 
 importer="texture"
 type="CompressedTexture2D"
-uid="uid://cljwckw7oogis"
-path="res://.godot/imported/playerGrey_walk2.png-34d2d916366100182d08037c51884043.ctex"
+uid="uid://cieqnir3bi2v6"
+path="res://.godot/imported/enemyFlyingAlt_2.png-411bb082dd6e825832cc003ae2f9f9d0.ctex"
 metadata={
 "vram_texture": false
 }
 
 [deps]
 
-source_file="res://art/playerGrey_walk2.png"
-dest_files=["res://.godot/imported/playerGrey_walk2.png-34d2d916366100182d08037c51884043.ctex"]
+source_file="res://characters/mob/sprite/enemyFlyingAlt_2.png"
+dest_files=["res://.godot/imported/enemyFlyingAlt_2.png-411bb082dd6e825832cc003ae2f9f9d0.ctex"]
 
 [params]
