+++ characters/mob/sprite/enemyWalking_1.png
