+++ characters/mob/sprite/enemyWalking_1.png.import
@@ -2,16 +2,16 @@
 
 importer="texture"
 type="CompressedTexture2D"
-uid="uid://ecwsoh0lu0a3"
-path="res://.godot/imported/enemySwimming_2.png-4c0cbc0732264c4ea3290340bd4a0a62.ctex"
+uid="uid://77yx8hnvverc"
+path="res://.godot/imported/enemyWalking_1.png-741e1f3c3bebaca6c4f84613c1e28ac2.ctex"
 metadata={
 "vram_texture": false
 }
 
 [deps]
 
-source_file="res://art/enemySwimming_2.png"
-dest_files=["res://.godot/imported/enemySwimming_2.png-4c0cbc0732264c4ea3290340bd4a0a62.ctex"]
+source_file="res://characters/mob/sprite/enemyWalking_1.png"
+dest_files=["res://.godot/imported/enemyWalking_1.png-741e1f3c3bebaca6c4f84613c1e28ac2.ctex"]
 
 [params]
