+++ characters/mob/sprite/enemyWalking_2.png
