+++ characters/mob/sprite/enemyWalking_2.png.import
@@ -2,16 +2,16 @@
 
 importer="texture"
 type="CompressedTexture2D"
-uid="uid://cfn8vahussm1g"
-path="res://.godot/imported/playerGrey_up2.png-d6aba85f5f2675ebc7045efa7552ee79.ctex"
+uid="uid://ch2x88ihemxbb"
+path="res://.godot/imported/enemyWalking_2.png-967722dc97b0f8816728afcb6730181c.ctex"
 metadata={
 "vram_texture": false
 }
 
 [deps]
 
-source_file="res://art/playerGrey_up2.png"
-dest_files=["res://.godot/imported/playerGrey_up2.png-d6aba85f5f2675ebc7045efa7552ee79.ctex"]
+source_file="res://characters/mob/sprite/enemyWalking_2.png"
+dest_files=["res://.godot/imported/enemyWalking_2.png-967722dc97b0f8816728afcb6730181c.ctex"]
 
 [params]
