+++ characters/player/player.tscn
@@ -1,0 +1,59 @@
+[gd_scene load_steps=9 format=3 uid="uid://f0qw34nkp1lv"]
+
+[ext_resource type="Texture2D" uid="uid://dbe8nb7563nc2" path="res://characters/player/sprite/playerGrey_up1.png" id="1_1a3f0"]
+[ext_resource type="Script" path="res://characters/player/player.cs" id="1_qufuk"]
+[ext_resource type="Texture2D" uid="uid://c7wsh4gjgris4" path="res://characters/player/sprite/playerGrey_dead1.png" id="2_nyice"]
+[ext_resource type="Texture2D" uid="uid://cfn8vahussm1g" path="res://characters/player/sprite/playerGrey_up2.png" id="2_plj0f"]
+[ext_resource type="Texture2D" uid="uid://cb0xvgdcg7b0" path="res://characters/player/sprite/playerGrey_walk1.png" id="3_1cdh6"]
+[ext_resource type="Texture2D" uid="uid://cljwckw7oogis" path="res://characters/player/sprite/playerGrey_walk2.png" id="4_x1yaj"]
+
+[sub_resource type="SpriteFrames" id="SpriteFrames_g6wpo"]
+animations = [{
+"frames": [{
+"duration": 1.0,
+"texture": ExtResource("2_nyice")
+}],
+"loop": true,
+"name": &"dead",
+"speed": 5.0
+}, {
+"frames": [{
+"duration": 1.0,
+"texture": ExtResource("1_1a3f0")
+}, {
+"duration": 1.0,
+"texture": ExtResource("2_plj0f")
+}],
+"loop": true,
+"name": &"up",
+"speed": 5.0
+}, {
+"frames": [{
+"duration": 1.0,
+"texture": ExtResource("3_1cdh6")
+}, {
+"duration": 1.0,
+"texture": ExtResource("4_x1yaj")
+}],
+"loop": true,
+"name": &"walk",
+"speed": 5.0
+}]
+
+[sub_resource type="CapsuleShape2D" id="CapsuleShape2D_bm22o"]
+radius = 27.0
+height = 68.0
+
+[node name="player" type="Area2D"]
+script = ExtResource("1_qufuk")
+
+[node name="AnimatedSprite2D" type="AnimatedSprite2D" parent="."]
+scale = Vector2(0.5, 0.5)
+sprite_frames = SubResource("SpriteFrames_g6wpo")
+animation = &"walk"
+frame_progress = 0.565453
+
+[node name="CollisionShape2D" type="CollisionShape2D" parent="."]
+shape = SubResource("CapsuleShape2D_bm22o")
+
+[connection signal="body_entered" from="." to="." method="OnBodyEntered"]
