+++ characters/player/sprite/playerGrey_up2.png
